Refactoring and fix tests
