+++ src/api/route-services/projects.js
@@ -12,6 +12,7 @@ const ExperimentService = require('./experiment');
 
 const samplesService = new SamplesService();
 const experimentService = new ExperimentService();
+
 class ProjectsService {
   constructor() {
     this.tableName = `projects-${config.clusterEnv}`;
@@ -210,39 +211,6 @@ class ProjectsService {
       if (e.statusCode === 400) throw new NotFoundError('Project not found');
       throw e;
     }
-  }
-
-  async getGem2sParams(experimentId) {
-    const experiment = await experimentService.getExperimentData(experimentId);
-    const { samples } = await samplesService.getSamplesByExperimentId(experimentId);
-    const { metadataKeys } = await this.getProject(experiment.projectId);
-
-    const defaultMetadataValue = 'N.A.';
-
-    const samplesEntries = Object.entries(samples);
-
-    const taskParams = {
-      projectId: experiment.projectId,
-      experimentName: experiment.experimentName,
-      organism: experiment.meta.organism,
-      input: { type: experiment.meta.type },
-      sampleIds: samplesEntries.map(([sampleId]) => sampleId),
-      sampleNames: samplesEntries.map(([, sample]) => sample.name),
-    };
-
-    if (metadataKeys.length) {
-      taskParams.metadata = metadataKeys.reduce((acc, key) => {
-        // Make sure the key does not contain '-' as it will cause failure in GEM2S
-        const sanitizedKey = key.replace(/-+/g, '_');
-
-        acc[sanitizedKey] = samplesEntries.map(
-          ([, sample]) => sample.metadata[key] || defaultMetadataValue,
-        );
-        return acc;
-      }, {});
-    }
-
-    return taskParams;
   }
 }
